refactor: modernize Python packaging and CI
- Replace setup.py/requirements.txt with pyproject.toml
- Rename package from 'dc' to 'dot_conf' for clarity
- Move source to src/dot_conf/ following best practices
- Add GitHub Actions workflow for CI
- Update README with new installation instructions
- Pin dependencies with version constraints
- Update test imports to use new package name
- Remove Travis CI configuration
