+++ pyproject.toml
@@ -1,0 +1,35 @@
+[build-system]
+requires = ["setuptools>=61.0"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "dot-conf"
+version = "0.0.3"
+authors = [
+  { name="Ben Zhang", email="benzhangniu@gmail.com" },
+]
+description = "automatically manage dotfiles"
+readme = "README.md"
+requires-python = ">=3.7"
+classifiers = [
+    "Programming Language :: Python :: 3",
+    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
+    "Operating System :: OS Independent",
+]
+dependencies = [
+    "strictyaml>=1.7.3,<1.8.0"
+]
+
+[project.urls]
+"Homepage" = "https://github.com/benzhang/dot-conf"
+
+[project.scripts]
+"dot-conf" = "dot_conf.__main__:main"
+
+[project.optional-dependencies]
+test = [
+    "pyfakefs>=5.9.1,<5.10.0",
+]
+
+[tool.setuptools.packages.find]
+where = ["src"]
