--- requirements.txt
@@ -1,3 +1,0 @@
-pathlib2
-pyfakefs
-strictyaml
